+++ src/main/java/es/princip/getp/domain/auth/application/AuthService.java
@@ -1,11 +1,11 @@
 package es.princip.getp.domain.auth.application;
 
 import es.princip.getp.domain.auth.domain.TokenVerification;
+import es.princip.getp.domain.auth.domain.TokenVerificationRepository;
 import es.princip.getp.domain.auth.dto.request.LoginRequest;
 import es.princip.getp.domain.auth.dto.response.Token;
 import es.princip.getp.domain.auth.exception.LoginErrorCode;
 import es.princip.getp.domain.auth.exception.TokenErrorCode;
-import es.princip.getp.domain.auth.repository.TokenVerificationRepository;
 import es.princip.getp.domain.member.application.MemberService;
 import es.princip.getp.infra.exception.BusinessLogicException;
 import es.princip.getp.infra.security.details.PrincipalDetails;
