+++ src/main/java/es/princip/getp/domain/auth/domain/EmailVerificationRepository.java
@@ -1,6 +1,5 @@
-package es.princip.getp.domain.auth.repository;
+package es.princip.getp.domain.auth.domain;
 
-import es.princip.getp.domain.auth.domain.EmailVerification;
 import org.springframework.data.keyvalue.repository.KeyValueRepository;
 
 public interface EmailVerificationRepository extends KeyValueRepository<EmailVerification, String> {
