+++ src/main/java/es/princip/getp/domain/people/application/PeopleLikeService.java
@@ -4,8 +4,8 @@ import es.princip.getp.domain.client.application.ClientService;
 import es.princip.getp.domain.client.domain.Client;
 import es.princip.getp.domain.people.domain.People;
 import es.princip.getp.domain.people.domain.PeopleLike;
+import es.princip.getp.domain.people.domain.PeopleLikeRepository;
 import es.princip.getp.domain.people.exception.PeopleLikeErrorCode;
-import es.princip.getp.domain.people.repository.PeopleLikeRepository;
 import es.princip.getp.infra.exception.BusinessLogicException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
