+++ src/main/java/es/princip/getp/domain/people/application/PeopleService.java
@@ -3,11 +3,11 @@ package es.princip.getp.domain.people.application;
 import es.princip.getp.domain.member.application.MemberService;
 import es.princip.getp.domain.member.dto.request.UpdateMemberRequest;
 import es.princip.getp.domain.people.domain.People;
+import es.princip.getp.domain.people.domain.PeopleRepository;
 import es.princip.getp.domain.people.dto.request.CreatePeopleRequest;
 import es.princip.getp.domain.people.dto.request.UpdatePeopleRequest;
 import es.princip.getp.domain.people.dto.response.people.CardPeopleResponse;
 import es.princip.getp.domain.people.exception.PeopleErrorCode;
-import es.princip.getp.domain.people.repository.PeopleRepository;
 import es.princip.getp.infra.exception.BusinessLogicException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
