+++ src/main/java/es/princip/getp/domain/people/domain/PeopleRepository.java
@@ -1,9 +1,10 @@
-package es.princip.getp.domain.people.repository;
+package es.princip.getp.domain.people.domain;
 
-import es.princip.getp.domain.people.domain.People;
-import java.util.Optional;
+import es.princip.getp.domain.people.infra.PeopleQueryDslRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
 
 @Repository
 public interface PeopleRepository extends JpaRepository<People, Long>, PeopleQueryDslRepository {
