+++ src/main/java/es/princip/getp/domain/people/infra/PeopleQueryDslRepository.java
@@ -1,8 +1,8 @@
-package es.princip.getp.domain.people.repository;
+package es.princip.getp.domain.people.infra;
 
+import es.princip.getp.domain.people.dto.response.people.CardPeopleResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import es.princip.getp.domain.people.dto.response.people.CardPeopleResponse;
 
 public interface PeopleQueryDslRepository {
