+++ src/main/java/es/princip/getp/domain/project/application/ProjectApplicationService.java
@@ -4,8 +4,8 @@ import es.princip.getp.domain.people.application.PeopleService;
 import es.princip.getp.domain.people.domain.People;
 import es.princip.getp.domain.project.domain.Project;
 import es.princip.getp.domain.project.domain.ProjectApplication;
+import es.princip.getp.domain.project.domain.ProjectApplicationRepository;
 import es.princip.getp.domain.project.dto.request.ApplicateProjectRequest;
-import es.princip.getp.domain.project.repository.ProjectApplicationRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
