+++ src/main/java/es/princip/getp/domain/project/application/ProjectLikeService.java
@@ -4,8 +4,8 @@ import es.princip.getp.domain.people.application.PeopleService;
 import es.princip.getp.domain.people.domain.People;
 import es.princip.getp.domain.project.domain.Project;
 import es.princip.getp.domain.project.domain.ProjectLike;
+import es.princip.getp.domain.project.domain.ProjectLikeRepository;
 import es.princip.getp.domain.project.exception.ProjectLikeErrorCode;
-import es.princip.getp.domain.project.repository.ProjectLikeRepository;
 import es.princip.getp.infra.exception.BusinessLogicException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
