+++ src/main/java/es/princip/getp/domain/project/application/ProjectService.java
@@ -3,9 +3,9 @@ package es.princip.getp.domain.project.application;
 import es.princip.getp.domain.client.application.ClientService;
 import es.princip.getp.domain.client.domain.Client;
 import es.princip.getp.domain.project.domain.Project;
+import es.princip.getp.domain.project.domain.ProjectRepository;
 import es.princip.getp.domain.project.dto.request.CreateProjectRequest;
 import es.princip.getp.domain.project.exception.ProjectErrorCode;
-import es.princip.getp.domain.project.repository.ProjectRepository;
 import es.princip.getp.infra.exception.BusinessLogicException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
