+++ src/main/java/es/princip/getp/domain/project/domain/ProjectApplicationRepository.java
@@ -1,6 +1,5 @@
-package es.princip.getp.domain.project.repository;
+package es.princip.getp.domain.project.domain;
 
-import es.princip.getp.domain.project.domain.ProjectApplication;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProjectApplicationRepository extends JpaRepository<ProjectApplication, Long> {
