+++ src/test/java/es/princip/getp/domain/member/application/MemberServiceTest.java
@@ -1,11 +1,11 @@
 package es.princip.getp.domain.member.application;
 
 import es.princip.getp.domain.member.domain.Member;
+import es.princip.getp.domain.member.domain.MemberRepository;
 import es.princip.getp.domain.member.domain.MemberType;
 import es.princip.getp.domain.member.dto.request.CreateMemberRequest;
 import es.princip.getp.domain.member.dto.request.UpdateMemberRequest;
 import es.princip.getp.domain.member.exception.MemberErrorCode;
-import es.princip.getp.domain.member.repository.MemberRepository;
 import es.princip.getp.domain.serviceTerm.application.ServiceTermService;
 import es.princip.getp.domain.serviceTerm.fixture.ServiceTermFixture;
 import es.princip.getp.domain.storage.application.ImageStorageService;
