+++ src/test/java/es/princip/getp/domain/people/application/PeopleLikeServiceTest.java
@@ -5,9 +5,9 @@ import es.princip.getp.domain.client.domain.Client;
 import es.princip.getp.domain.member.domain.Member;
 import es.princip.getp.domain.people.domain.People;
 import es.princip.getp.domain.people.domain.PeopleLike;
+import es.princip.getp.domain.people.domain.PeopleLikeRepository;
 import es.princip.getp.domain.people.exception.PeopleErrorCode;
 import es.princip.getp.domain.people.exception.PeopleLikeErrorCode;
-import es.princip.getp.domain.people.repository.PeopleLikeRepository;
 import es.princip.getp.infra.exception.BusinessLogicException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
