+++ src/test/java/es/princip/getp/domain/people/application/PeopleProfileServiceTest.java
@@ -3,11 +3,11 @@ package es.princip.getp.domain.people.application;
 import es.princip.getp.domain.member.domain.Member;
 import es.princip.getp.domain.people.domain.People;
 import es.princip.getp.domain.people.domain.PeopleProfile;
+import es.princip.getp.domain.people.domain.PeopleProfileRepository;
+import es.princip.getp.domain.people.domain.PeopleRepository;
 import es.princip.getp.domain.people.dto.request.CreatePeopleProfileRequest;
 import es.princip.getp.domain.people.dto.request.UpdatePeopleProfileRequest;
 import es.princip.getp.domain.people.fixture.PeopleProfileFixture;
-import es.princip.getp.domain.people.repository.PeopleProfileRepository;
-import es.princip.getp.domain.people.repository.PeopleRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
