+++ src/test/java/es/princip/getp/domain/people/application/PeopleServiceTest.java
@@ -4,10 +4,10 @@ import es.princip.getp.domain.member.application.MemberService;
 import es.princip.getp.domain.member.domain.Member;
 import es.princip.getp.domain.member.dto.request.UpdateMemberRequest;
 import es.princip.getp.domain.people.domain.People;
+import es.princip.getp.domain.people.domain.PeopleRepository;
 import es.princip.getp.domain.people.dto.request.CreatePeopleRequest;
 import es.princip.getp.domain.people.dto.request.UpdatePeopleRequest;
 import es.princip.getp.domain.people.dto.response.people.CardPeopleResponse;
-import es.princip.getp.domain.people.repository.PeopleRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
