+++ src/test/java/es/princip/getp/domain/project/application/ProjectApplicationServiceTest.java
@@ -7,8 +7,8 @@ import es.princip.getp.domain.people.application.PeopleService;
 import es.princip.getp.domain.people.domain.People;
 import es.princip.getp.domain.project.domain.Project;
 import es.princip.getp.domain.project.domain.ProjectApplication;
+import es.princip.getp.domain.project.domain.ProjectApplicationRepository;
 import es.princip.getp.domain.project.dto.request.ApplicateProjectRequest;
-import es.princip.getp.domain.project.repository.ProjectApplicationRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
