+++ src/test/java/es/princip/getp/domain/project/application/ProjectLikeServiceTest.java
@@ -4,9 +4,9 @@ import es.princip.getp.domain.member.domain.Member;
 import es.princip.getp.domain.people.application.PeopleService;
 import es.princip.getp.domain.people.domain.People;
 import es.princip.getp.domain.project.domain.Project;
+import es.princip.getp.domain.project.domain.ProjectLikeRepository;
 import es.princip.getp.domain.project.exception.ProjectErrorCode;
 import es.princip.getp.domain.project.exception.ProjectLikeErrorCode;
-import es.princip.getp.domain.project.repository.ProjectLikeRepository;
 import es.princip.getp.infra.exception.BusinessLogicException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
