+++ src/test/java/es/princip/getp/domain/project/application/ProjectServiceTest.java
@@ -6,9 +6,9 @@ import es.princip.getp.domain.client.fixture.ClientFixture;
 import es.princip.getp.domain.member.domain.Member;
 import es.princip.getp.domain.member.fixture.MemberFixture;
 import es.princip.getp.domain.project.domain.Project;
+import es.princip.getp.domain.project.domain.ProjectRepository;
 import es.princip.getp.domain.project.dto.request.CreateProjectRequest;
 import es.princip.getp.domain.project.fixture.ProjectFixture;
-import es.princip.getp.domain.project.repository.ProjectRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
